+++ src/careamics/config/builder/mixins/data_params.py
@@ -1,5 +1,5 @@
 from collections.abc import Sequence
-from typing import Any, Literal, TypeVar, overload
+from typing import Any, Literal, overload
 
 from careamics.config.augmentations import (
     XYFlipConfig,
@@ -8,44 +8,11 @@ from careamics.config.augmentations import (
 from careamics.config.factories.data_factory import (
     list_spatial_augmentations,
 )
-from careamics.config.factories.training_factory import update_trainer_params
 
-from .config_builder import ConfigBuilder
-
-ConfigBuilderT = TypeVar("ConfigBuilderT", bound=ConfigBuilder)
+from ..config_builder import ConfigBuilder, ConfigBuilderT
 
 
-class TrainingParamMixin:
-    def __init__(self: ConfigBuilder, *args: Any, **kwargs: Any):
-        super().__init__(*args, **kwargs)
-
-        self.config_dict["training_config"].setdefault(
-            "trainer_params", update_trainer_params({}, self.num_epochs, self.num_steps)
-        )
-
-    def set_trainer_params(self: ConfigBuilderT, **kwargs: Any) -> ConfigBuilderT:
-        assert isinstance(self.config_dict["training_config"]["trainer_params"], dict)
-        self.config_dict["training_config"]["trainer_params"].update(kwargs)
-        return self
-
-    def set_checkpoint_params(self: ConfigBuilderT, **kwargs: Any) -> ConfigBuilderT:
-        self.config_dict["training_config"]["checkpoint_params"] = kwargs
-        return self
-
-    def set_early_stopping_params(
-        self: ConfigBuilderT, **kwargs: Any
-    ) -> ConfigBuilderT:
-        self.config_dict["training_config"]["early_stopping_params"] = kwargs
-        return self
-
-    def set_logger(
-        self: ConfigBuilderT, name: Literal["wandb", "tensorboard"] | None = None
-    ) -> ConfigBuilderT:
-        self.config_dict["training_config"]["logger"] = name
-        return self
-
-
-class DataParamMixin:
+class DataParamsMixin:
     def __init__(self: ConfigBuilder, *args: Any, **kwargs: Any):
         super().__init__(*args, **kwargs)
 
@@ -236,90 +203,4 @@ class DataParamMixin:
             **filter_config_dict,
             **kwargs,
         }
-        return self
-
-
-class OptimizerParamMixin:
-    def __init__(self: ConfigBuilder, *args: Any, **kwargs: Any):
-        super().__init__(*args, **kwargs)
-
-    def set_optimizer(
-        self: ConfigBuilderT, name: Literal["Adam", "Adamax", "SGD"], **kwargs: Any
-    ) -> ConfigBuilderT:
-        self.config_dict["algorithm_config"]["optimizer"] = {
-            "name": name,
-            "parameters": kwargs,
-        }
-        return self
-
-    def set_lr_scheduler(
-        self: ConfigBuilderT,
-        name: Literal["ReduceLROnPlateau", "StepLR"],
-        **kwargs: Any,
-    ) -> ConfigBuilderT:
-        self.config_dict["algorithm_config"]["lr_scheduler"] = {
-            "name": name,
-            "parameters": kwargs,
-        }
-        return self
-
-
-class UnetParamMixin:
-    def __init__(self: ConfigBuilder, *args: Any, **kwargs: Any):
-        super().__init__(*args, **kwargs)
-
-        model_config: dict[str, Any] = {
-            "conv_dims": 3 if self.is_3D else 2,
-        }
-        if self.n_channels_in is not None:
-            model_config["in_channels"] = self.n_channels_in
-        if self.n_channels_out is not None:
-            model_config["num_classes"] = self.n_channels_out
-
-        self.config_dict["algorithm_config"].setdefault("model", {})
-        assert isinstance(self.config_dict["algorithm_config"]["model"], dict)
-
-        # TODO: error if architecture already exists in data_config?
-        self.config_dict["algorithm_config"]["model"]["architecture"] = "UNet"
-        for key, value in model_config.items():
-            self.config_dict["algorithm_config"]["model"].setdefault(key, value)
-
-    def set_model_params(
-        self: ConfigBuilderT,
-        independent_channels: bool | None = None,
-        depth: int | None = None,
-        num_channels_init: int | None = None,
-        residual: bool | None = None,
-        final_activation: (
-            Literal["None", "Sigmoid", "Softmax", "Tanh", "ReLU", "LeakyReLU"] | None
-        ) = None,
-        use_batch_norm: bool | None = None,
-    ) -> ConfigBuilderT:
-        assert isinstance(self.config_dict["algorithm_config"]["model"], dict)
-
-        if independent_channels is not None:
-            self.config_dict["algorithm_config"]["model"][
-                "independent_channels"
-            ] = independent_channels
-
-        if depth is not None:
-            self.config_dict["algorithm_config"]["model"]["depth"] = depth
-
-        if num_channels_init is not None:
-            self.config_dict["algorithm_config"]["model"][
-                "num_channels_init"
-            ] = num_channels_init
-
-        if residual is not None:
-            self.config_dict["algorithm_config"]["model"]["residual"] = residual
-
-        if final_activation is not None:
-            self.config_dict["algorithm_config"]["model"][
-                "final_activation"
-            ] = final_activation
-
-        if use_batch_norm is not None:
-            self.config_dict["algorithm_config"]["model"][
-                "use_batch_norm"
-            ] = use_batch_norm
         return self
